+++ test/resources/unused_files_analyzer/unused_files_example.dart
@@ -1,5 +1,8 @@
 import 'package:dart_code_metrics/config.dart';
 
+import 'unconditional_file.dart' if (dart.library.html) 'conditional_file.dart'
+    as file;
+
 import 'imported_file.dart';
 
 export 'exported_file.dart';